*: move internal packages to lib/internal
