+++ lib/internal/tarball/walk.go
